backend/requirements.txt
fastapi 
uvicorn 
pdfplumber 
python-multipart 
supabase 
python-dotenv
groq